Fixed issues with multi-threading in GameLoop.
There was noticeable graphical glitches when the bike was moving
really fast. This was caused by the gameUpdate() being called
in parallel to the gameDraw(), which meant that some draw calls
got an older bike position whereas others got a new one. A
lock has been introduced to fix this problem.